@@ -1,47 +1,37 @@
 {
-  "name": "client",
+  "name": "client2",
   "version": "0.1.0",
   "private": true,
   "scripts": {
     "serve": "vue-cli-service serve",
     "build": "vue-cli-service build",
-    "test:unit": "vue-cli-service test:unit",
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "axios": "^0.21.3",
-    "bootstrap": "^4.0.0",
+    "@popperjs/core": "^2.10.2",
+    "axios": "^0.24.0",
+    "bootstrap": "^5.0.0-beta3",
     "bootstrap-vue": "^2.21.2",
     "core-js": "^3.6.5",
-    "dotenv": "^10.0.0",
-    "jquery": "^3.6.0",
-    "popper.js": "^1.16.1",
-    "vue": "^2.6.11",
-    "vue-axios": "^3.2.5",
+    "sass-loader": "^12.2.0",
+    "vue": "^3.0.0",
+    "vue-axios": "^3.3.7",
     "vue-cookie": "^1.1.4",
-    "vue-cookies": "^1.7.4",
-    "vue-router": "^3.2.0",
-    "vuex": "^3.4.0"
+    "vue-router": "^4.0.0-0",
+    "vuex": "^4.0.0-0"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~4.5.0",
     "@vue/cli-plugin-eslint": "~4.5.0",
     "@vue/cli-plugin-router": "~4.5.0",
-    "@vue/cli-plugin-unit-jest": "~4.5.0",
     "@vue/cli-plugin-vuex": "~4.5.0",
     "@vue/cli-service": "~4.5.0",
+    "@vue/compiler-sfc": "^3.0.0",
     "@vue/eslint-config-prettier": "^6.0.0",
-    "@vue/test-utils": "^1.0.3",
     "babel-eslint": "^10.1.0",
-    "bootstrap": "^4.0.0-beta",
     "eslint": "^6.7.2",
     "eslint-plugin-prettier": "^3.3.1",
-    "eslint-plugin-vue": "^6.2.2",
-    "jquery": "^3.6.0",
-    "node-sass": "^4.12.0",
-    "popper.js": "^1.16.1",
-    "prettier": "^2.2.1",
-    "sass-loader": "^8.0.2",
-    "vue-template-compiler": "^2.6.11"
+    "eslint-plugin-vue": "^7.0.0",
+    "prettier": "^2.2.1"
   }
 }
