@@ -1,48 +1,9 @@
-import Vue from "vue";
+import { createApp } from "vue";
 import App from "./App.vue";
 import router from "./router";
 import store from "./store";
 
-import "bootstrap";
 import "bootstrap/dist/css/bootstrap.min.css";
+import "bootstrap";
 
-Vue.config.productionTip = false;
-
-// router guards dones not redirect well but it work
-router.beforeEach((to, from, next) => {
-  if (to.matched.some((record) => record.meta.requiresAuth)) {
-    if (!store.getters.loggedIn) {
-      next({
-        name: "Login",
-      });
-    } else {
-      next();
-    }
-  } else if (to.matched.some((record) => record.meta.requiresVisitor)) {
-    if (store.getters.loggedIn) {
-      next({
-        name: "Home",
-      });
-    } else {
-      next();
-    }
-  } else {
-    next();
-  }
-});
-
-new Vue({
-  router,
-  store,
-  render: (h) => h(App),
-}).$mount("#app");
-
-// Does it help?
-const allowCrossDomain = function (req, res, next) {
-  res.header("Access-Control-Allow-Origin", "*");
-  res.header("Access-Control-Allow-Methods", "*");
-  res.header("Access-Control-Allow-Headers", "*");
-  next();
-};
-
-Vue.use(allowCrossDomain);
+createApp(App).use(store).use(router).mount("#app");
