@@ -1,30 +1,8 @@
-import Vue from "vue";
-import Vuex from "vuex";
-import VueCookies from "vue-cookies";
-Vue.use(VueCookies);
-Vue.use(Vuex);
+import { createStore } from "vuex";
 
-export default new Vuex.Store({
-  state: {
-    token: localStorage.getItem("access_token") || null,
-    // access_token_cookie: Vue.$cookies.get("access_token_cookie") || null,
-    api_url: "http://127.0.0.1:8080",
-  },
-  mutations: {
-    updateToken(state, token) {
-      state.token = token;
-      // state.access_token_cookie = token;
-    },
-    distroyToken(state) {
-      state.token = null;
-      // state.access_token_cookie = null;
-    },
-  },
+export default createStore({
+  state: {},
+  mutations: {},
   actions: {},
   modules: {},
-  getters: {
-    loggedIn(state) {
-      return state.token != null;
-    },
-  },
 });
