@@ -8,6 +8,7 @@
         v-model="newTodo"
         @keyup.enter="addTodo"
       />
+
       <!-- List -->
       <div class="container">
         <div class="todo-list">
@@ -20,7 +21,7 @@
               />
               <div
                 v-if="!todo.editing"
-                @dblclick="editTodo(todo, index)"
+                @dblclick="editTodo(todo)"
                 class="todo-item-label"
                 :class="{ complete: todo.complete }"
               >
@@ -49,32 +50,16 @@
 </template>
 
 <script>
-import Vue from "vue";
 import axios from "axios";
-import VueAxios from "vue-axios";
-import VueCookie from "vue-cookie";
-
-Vue.use(VueCookie);
-Vue.use(VueAxios, axios);
 
 export default {
   name: "Todo",
-  beforeCreate() {
-    var token = this.$cookie.get("access_token_cookie");
-    console.log(token);
-    // add to local storage
-    localStorage.setItem("access_token", token); // token stores in cookies
-
-    // update the store state token
-    this.$store.commit("updateToken", token);
-  },
-
   data() {
     return {
       newTodo: "",
       idForTodo: "",
       todos: [],
-      baseUrl: this.$store.state.api_url + "/task",
+      baseUrl: "http://localhost:8080/tasks",
     };
   },
 
@@ -87,33 +72,19 @@ export default {
   },
 
   created() {
-    // this.loadToken();
     this.getTodos();
   },
 
   methods: {
-    loadToken() {
-      var token = this.$cookie.get("access_token_cookie");
-      console.log(token);
-      // add to local storage
-      localStorage.setItem("access_token", token); // token stores in cookies
-
-      // update the store state token
-      this.$store.commit("updateToken", token);
-    },
-
     getTodos() {
       const path = this.baseUrl;
-      axios.defaults.headers.common["Authorization"] =
-        "Bearer " + this.$store.state.token;
-
-      Vue.axios.get(path).then((res) => {
+      axios.get(path).then((res) => {
         res.data["tasks"].forEach((todo) => {
           this.todos.push({
-            id: todo._id,
-            text: todo.text,
-            complete: todo.complete,
-            author: todo.author,
+            id: todo.ID,
+            text: todo.Text,
+            complete: todo.Completed,
+            author: todo.User,
             editing: false,
           });
         });
@@ -128,12 +99,13 @@ export default {
       const path = this.baseUrl;
       var newEntry = {
         text: this.newTodo,
-        complete: false,
-        editing: false,
+        // complete: false,
+        // editing: false,
       };
-      Vue.axios
+
+      axios
         .post(path, newEntry)
-        .then((response) => (this.idForTodo = response.data.id))
+        // .then((response) => (this.idForTodo = response.data.id))
         .then(this.todos.push(newEntry));
       this.newTodo = "";
     },
@@ -141,9 +113,8 @@ export default {
     removeTodo(index) {
       var task_id = this.todos[index]["id"];
       this.todos.splice(index, 1);
-      // console.log(task_id);
       const path = this.baseUrl + "/" + task_id;
-      Vue.axios.delete(path);
+      axios.delete(path);
     },
 
     editTodo(todo) {
@@ -153,7 +124,7 @@ export default {
     updateTodo(todo, index) {
       var task_id = this.todos[index]["id"];
       const path = this.baseUrl + "/" + task_id;
-      Vue.axios.put(path, { text: todo.text });
+      axios.put(path, { text: todo.text });
     },
 
     doneEdit(todo, index) {
@@ -164,7 +135,7 @@ export default {
     updateStatus(todo, index) {
       var task_id = this.todos[index]["id"];
       const path = this.baseUrl + "/" + task_id + "/check";
-      Vue.axios.put(path, { complete: !todo.complete }); // i think it takes the value before changing so i NOT it
+      axios.put(path);
     },
   },
 };
@@ -182,7 +153,6 @@ export default {
 }
 
 .todo-item {
-  margin-bottom: 12px;
   display: flex;
   align-items: center;
   justify-content: space-between;
