@@ -1,5 +1,5 @@
 module.exports = {
     devServer: {
-          proxy: 'https://127.0.0.1:5000',
-      }
-  }
+        proxy: 'http://localhost:5000',
+    }
+}
